Move cli/preference/registry/util.go > registry/utils.go
