+++ pkg/odo/cli/preference/remove/registry.go
@@ -12,10 +12,10 @@ import (
 	ktemplates "k8s.io/kubectl/pkg/util/templates"
 
 	// odo packages
-	registryUtil "github.com/redhat-developer/odo/pkg/odo/cli/preference/registry/util"
 	"github.com/redhat-developer/odo/pkg/odo/cmdline"
 	"github.com/redhat-developer/odo/pkg/odo/genericclioptions"
 	"github.com/redhat-developer/odo/pkg/odo/genericclioptions/clientset"
+	registryUtil "github.com/redhat-developer/odo/pkg/registry"
 )
 
 const registryCommandName = "registry"
